fix: rename case sensitive name from transaction component
